+++ extensions/ql-vscode/src/contextual/fileRangeFromURI.ts
@@ -2,7 +2,7 @@ import * as vscode from "vscode";
 
 import { UrlValue, LineColumnLocation } from "../pure/bqrs-cli-types";
 import { isEmptyPath } from "../pure/bqrs-utils";
-import { DatabaseItem } from "../databases";
+import { DatabaseItem } from "../local-databases";
 
 export default function fileRangeFromURI(
   uri: UrlValue | undefined,
