+++ extensions/ql-vscode/src/databaseFetcher.ts
@@ -16,7 +16,7 @@ import { basename, join } from "path";
 import * as Octokit from "@octokit/rest";
 import { retry } from "@octokit/plugin-retry";
 
-import { DatabaseManager, DatabaseItem } from "./databases";
+import { DatabaseManager, DatabaseItem } from "./local-databases";
 import { showAndLogInformationMessage, tmpDir } from "./helpers";
 import { reportStreamProgress, ProgressCallback } from "./commandRunner";
 import { extLogger } from "./common";
