+++ extensions/ql-vscode/src/legacy-query-server/upgrades.ts
@@ -10,7 +10,7 @@ import * as messages from "../pure/legacy-messages";
 import * as qsClient from "./queryserver-client";
 import * as tmp from "tmp-promise";
 import { dirname } from "path";
-import { DatabaseItem } from "../databases";
+import { DatabaseItem } from "../local-databases";
 import { asError, getErrorMessage } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
