+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -40,7 +40,7 @@ import {
   getQueryText,
   QueryHistoryInfo,
 } from "./query-history-info";
-import { DatabaseManager } from "../databases";
+import { DatabaseManager } from "../local-databases";
 import { registerQueryHistoryScrubber } from "./query-history-scrubber";
 import {
   QueryStatus,
