+++ extensions/ql-vscode/src/query-server/query-runner.ts
@@ -1,6 +1,6 @@
 import { CancellationToken } from "vscode";
 import { ProgressCallback, UserCancellationException } from "../commandRunner";
-import { DatabaseItem } from "../databases";
+import { DatabaseItem } from "../local-databases";
 import {
   clearCache,
   ClearCacheParams,
