+++ extensions/ql-vscode/src/queryRunner.ts
@@ -1,7 +1,7 @@
 import { CancellationToken } from "vscode";
 import { CodeQLCliServer } from "./cli";
 import { ProgressCallback } from "./commandRunner";
-import { DatabaseItem } from "./databases";
+import { DatabaseItem } from "./local-databases";
 import { InitialQueryInfo, LocalQueryInfo } from "./query-results";
 import { QueryWithResults } from "./run-queries-shared";
