+++ extensions/ql-vscode/test/vscode-tests/cli-integration/databaseFetcher.test.ts
@@ -3,7 +3,7 @@ import { extensions, CancellationToken, Uri, window } from "vscode";
 
 import { CodeQLExtensionInterface } from "../../../src/extension";
 import { CodeQLCliServer } from "../../../src/cli";
-import { DatabaseManager } from "../../../src/databases";
+import { DatabaseManager } from "../../../src/local-databases";
 import {
   importArchiveDatabase,
   promptImportInternetDatabase,
