+++ extensions/ql-vscode/test/vscode-tests/cli-integration/queries.test.ts
@@ -15,7 +15,7 @@ import {
 } from "fs-extra";
 import { load, dump } from "js-yaml";
 
-import { DatabaseItem, DatabaseManager } from "../../../src/databases";
+import { DatabaseItem, DatabaseManager } from "../../../src/local-databases";
 import { CodeQLExtensionInterface } from "../../../src/extension";
 import { cleanDatabases, dbLoc, storagePath } from "../global.helper";
 import { importArchiveDatabase } from "../../../src/databaseFetcher";
