+++ extensions/ql-vscode/test/vscode-tests/global.helper.ts
@@ -2,7 +2,7 @@ import { join } from "path";
 import { load, dump } from "js-yaml";
 import { realpathSync, readFileSync, writeFileSync } from "fs-extra";
 import { commands } from "vscode";
-import { DatabaseManager } from "../../src/databases";
+import { DatabaseManager } from "../../src/local-databases";
 import { CodeQLCliServer } from "../../src/cli";
 import { removeWorkspaceRefs } from "../../src/variant-analysis/run-remote-query";
