+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/databases.test.ts
@@ -11,7 +11,7 @@ import {
   FullDatabaseOptions,
   findSourceArchive,
   DatabaseResolver,
-} from "../../../src/databases";
+} from "../../../src/local-databases";
 import { Logger } from "../../../src/common";
 import { ProgressCallback } from "../../../src/commandRunner";
 import { CodeQLCliServer, DbInfo } from "../../../src/cli";
