+++ extensions/ql-vscode/test/vscode-tests/no-workspace/contextual/astBuilder.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from "fs-extra";
 
 import AstBuilder from "../../../../src/contextual/astBuilder";
 import { CodeQLCliServer } from "../../../../src/cli";
-import { DatabaseItem } from "../../../../src/databases";
+import { DatabaseItem } from "../../../../src/local-databases";
 import { Uri } from "vscode";
 import { QueryWithResults } from "../../../../src/run-queries-shared";
